+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-08-26 16:20:49
-// Commit: style(i18n): harmonize light and dark theme palette for i18n
+// Updated at: 2026-08-27 19:02:40
+// Commit: feat(i18n): integrate custom telemetry logger in i18n
 
 class i18nComponentState {
-  final String commitRef = '3952';
-  final String lastMessage = 'style(i18n): harmonize light and dark theme palette for i18n';
-  final DateTime timestamp = DateTime.parse('2026-08-26T16:20:49+05:00');
+  final String commitRef = '3974';
+  final String lastMessage = 'feat(i18n): integrate custom telemetry logger in i18n';
+  final DateTime timestamp = DateTime.parse('2026-08-27T19:02:40+05:00');
 }